+++ src/main/java/com/nexorithmm/java/Temp.java
@@ -1,0 +1,4 @@
+package com.nexorithmm.java;
+
+public class Temp {
+}