+++ src/main/java/com/nexorithmm/java/basics/FirstJavaProgram.java
@@ -1,0 +1,21 @@
+package com.nexorithmm.java.basics;
+
+public class FirstJavaProgram {
+    public static void main(String[] args) {
+
+        System.out.println("Hello, Nexorithmm!");
+
+        System.out.println("My name is Arjun Sarkar");
+
+        System.out.println("I am learning Java.");
+
+        System.out.println("My Goal is to become a Java Full Stack Developer.");
+
+        System.out.print("Your College Name");
+
+        System.out.print("Your department");
+
+        System.out.println("Your Name");
+
+    }
+}